Add paginated news detail view with pager
